add a simple quad tree implementation

+++ simulator-gui/src/main/scala/it/unibo/scafi/simulation/gui/launcher/scalaFX/WorldConfig.scala
@@ -70,7 +70,7 @@ object WorldConfig {
     world ++ node
   }
 
-  def gridLike2D(row : Int, column : Int, distance : Int): Unit = {
+  def gridLike2D(row : Int, column : Int, distance : Double): Unit = {
     var node : Set[NODE] = Set()
     val z = 0
     var nodes = 0;
